==> AgingFlyCellAtlas_KEGG.pathway_analysis/Fruitfly-GEM_metabolite_neighbors/00_dataset_preprocessing.R <==
library(Matrix)
#https://stackoverflow.com/questions/51467276/how-to-find-the-column-means-for-a-sparse-matrix-excluding-0-values
# Note: rows or columsn with all 0 are dropped
colMeans_drop0 <- function (dgCMat) {
  nnz_per_col <- diff(dgCMat@p)
  ColInd <- rep.int(1:ncol(dgCMat), nnz_per_col)
  sapply(split(dgCMat@x, ColInd), mean)
}
rowMeans_drop0 <- function (dgCMat) {
  RowInd <- dgCMat@i + 1
  sapply(split(dgCMat@x, RowInd), mean)
}
# number of non-zero: https://stackoverflow.com/questions/51560456/r-package-matrix-get-number-of-non-zero-entries-per-rows-columns-of-a-sparse
##########
library(SingleCellExperiment)
library(zellkonverter)
library(scater);library(scran)
library(ggplot2);library(gridExtra);
library(tidyverse);
library(SummarizedExperiment)

###############################################################
## read in kegg module info
gene.meta=readRDS('~/Documents/Data_AgingFlyCellAtlas/AFCA_gene.id.rds')
#gene.meta=readRDS('~/Documents/Data_AgingFlyCellAtlas/AFCA_gene.meta.rds')
gene.meta=gene.meta[gene.meta$FLYBASE!='',]
head(gene.meta)
dup.names=names(which(table(gene.meta$original.id)>1))

###############################################################
## read in data
if(F){ #only run once
  sce0=readH5AD('~/Documents/Data_AgingFlyCellAtlas/adata_headBody_S_v1.0.h5ad') # 15992  566254  
  #sce0=readH5AD('~/Documents/Data_AgingFlyCellAtlas/adata_head_S_v1.0.h5ad') # 15992 289981 
  #sce0=readH5AD('~/Documents/Data_AgingFlyCellAtlas/adata_body_S_v1.0.h5ad') # 15992 276273 
  sce0 #SingleCellExperiment 
  summary(sce0$n_genes_by_counts) 
  summary(sce0$total_counts)
  summary(sce0$pct_counts_mt) #all below 5%
  table(sce0$tissue)
  sce0=sce0[,-grep('unannotated',ignore.case = T,sce0$afca_annotation)] #remove the 'unannotation' 
  
  sce=sce0[,sce0$sex=='female' & sce0$n_genes_by_counts>=200 & sce0$total_counts>=1000 & sce0$pct_counts_mt<5]
  sce$afca_annotation=droplevels(sce$afca_annotation)
  sce$cell.type=sce$afca_annotation;
  sce$tc=paste(sce$tissue,sce$cell.type,sep=';')
  cell.meta=colData(sce)
  #saveRDS(cell.meta,'female_filtered.cell.meta.rds')
  #saveRDS(cell.meta,'female_head_filtered.cell.meta.rds')
  #saveRDS(cell.meta,'female_body_filtered.cell.meta.rds')
  
  #cell.meta$tc=paste(cell.meta$tissue,cell.meta$afca_annotation,sep=';')
  length(table(cell.meta$cell.type))  #152
  
  # select tc with >=100 cell in all 4 ages
  x=as.data.frame(cell.meta) %>% group_by(tissue,age,cell.type) %>% summarise(ncell=n()) %>% filter(ncell>=100)
  x1=x %>% spread(age,ncell)
  dim(x1) #63
  nage=apply(x1,1,function(i) sum(!is.na(i[-c(1,2)])) )
  pick.cell.type=x1[nage==4,]
  dim(pick.cell.type) #27 (16 in head or 11 in body)
  
  pick.cell.type$tc=paste(pick.cell.type$tissue,pick.cell.type$cell.type,sep=';')
  
  keep.sce=sce[,sce$tc %in% pick.cell.type$tc]
  writeH5AD(keep.sce,'sce_filteredBy_ncell100.h5ad')
}
